Avoid regenerating the same list over and over again
When polling the search API, facia tool regenerates the entire list of articles in the latest section.
However most of the times nothing changes.

This PR takes advantage of knockout observable and generates new articles only when necessary.

This came up investigating a performance issue.

Bonus item, when a new item appers, it flashes for a second, kinda like hack day winner...

For the moment it's behind a URL switch, I'd like to get some feedback first.